Extract GameObjectFactory interface
- Use a common interface for TileFactory and PlayerFactory

Explanation:

- After the last refactoring, there was some duplicate code in the createAndPlaceTile() and createAndPlacePlayer() methods.
- This is always an indicator that we might violate the DRY (Don't Repeat Yourself) principle.
  - However, there are two kinds of code duplication: duplication of behavior and *incidental* duplication. The latter means it looks like duplicate code syntax-wise, but it actually does something else. In this case we leave it alone.
  - But in our case it's the typical kind of code duplication that we shouldn't tolerate. The last two lines of the functions are identical completely identical and are doing the same thing. We could have easily pulled them out into their own method and invoke it from both places.
  - However, also the first line of the methods is basically doing the same thing - creating a GameObject (you can, in fact, just replace the Tile and Player type with GameObject). The only thing that keeps them from being completely identical is the type of the factories.
- Since both factories are using essentially the same interface to create game objects, the create(<enum-type>) method, we can just extract this common interface as a Java interface and make it official: they are both GameObjectFactory's.
- Now we can pull both implementations together in the new createAndPlaceGameObject() method.

Q&A:

Q: Why not just leave it alone? It's just two lines...
A: If in some future version of the game engine the way how objects are positioned is changed, we need to adapt this code in multiple places.
   It's very easy (and common!) to forget some of them. Also imagine we may add another kind of GameObject (for example: Predator).

Q: What on earth is this '<E extends Enum<E>>' thing!?
A: It's a way to declare a generic parameter with certain type restrictions. In this case we say whenever type 'E' is used, it must be an Enum.

Q: What does DRY mean?
A: It means Don't Repeat Yourself. Basically: copy and paste is the root of evil in any code base. Whenever you find yourself being left with duplicate code, try hard to refactor your code base.

Q: If there is DRY code, is there also WET code?
A: Yes, there is. Code is WET, When Everything is Terrible :-)